Add files via upload
This is the raw data on die numbers. The columns on the left side of the table contain direct information from the die studies that have been digitized. The right-hand column---beginning with _649_625---contain the number of dies for each 25-year range. In this case, it would be 649-625 BCE. The vast majority contain 0, especially as we pass into the Hellenistic and Roman periods, which have not yet been digitized. These have been filled by working with the column 'number_of_obverse_dies', 'date_range_min' and 'date_range_max'. Each date range column takes the difference between min and max dates and calculates the number that fall each date range. It is an estimate, based on the assumption that dies within a range may fall equally into any part of that range: this is likely untrue, but is a useful way of dealing with this information in rough strokes.