Add files via upload
This is the python code that creates the map visualization. This was done prior by CartoDB (and so relied on data stored in a SQL database). This was in the end not really needed (and also, my student subscription to Carto finally ran out and they wanted too much money). So I replaced it with Python producing a Leaflet map.